# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AirFlex API",
    "version": "1.0.0",
    "description": "Open-source P2P Airtime & Data Exchange Marketplace backed by the Stellar Network. All protected endpoints require a Bearer JWT obtained from POST /api/v1/auth/verify-otp.",
    "contact": {
      "name": "AirFlex Core Team",
      "url": "https://github.com/arflexx/Airflex"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT issued by POST /api/v1/auth/verify-otp. Pass as `Authorization: Bearer <token>`."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "Descriptive error message"
          }
        }
      },
      "ValidationErrorItem": {
        "type": "object",
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string",
            "example": "phone"
          },
          "message": {
            "type": "string",
            "example": "phone is required"
          }
        }
      },
      "ValidationErrorResponse": {
        "type": "object",
        "required": [
          "errors"
        ],
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationErrorItem"
            }
          }
        }
      },
      "PaginationMeta": {
        "type": "object",
        "required": [
          "page",
          "limit",
          "total",
          "totalPages"
        ],
        "properties": {
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 20
          },
          "total": {
            "type": "integer",
            "example": 100
          },
          "totalPages": {
            "type": "integer",
            "example": 5
          }
        }
      },
...
      "TradeStatus": {
        "type": "string",
        "enum": [
          "Active",
          "Locked",
          "Completed",
          "Cancelled",
          "Disputed"
        ]
      },
      "TradeOffer": {
        "type": "object",
        "required": [
          "id",
          "asset_type",
          "amount",
          "status",
          "expires_at",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "seller_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "Seller UUID. Present on detail/create responses but omitted from the public listing feed, which exposes `seller_handle` instead (issue #330)."
          },
          "seller_handle": {
            "type": "string",
            "nullable": true,
            "example": "@airflex_2f9a",
            "description": "Opaque, deterministic seller label derived from the user id. Included on the public listing feed instead of `seller_id` (issue #330)."
          },
          "buyer_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "nullable": true
          },
          "asset_type": {
            "type": "string",
            "example": "MTN_AIRTIME"
          },
          "amount": {
            "type": "number",
            "example": 500
          },
          "status": {
            "$ref": "#/components/schemas/TradeStatus"
          },
          "contract_listing_id": {
            "type": [
              "string",
              "null"
            ],
            "nullable": true,
            "example": "42"
          },
          "escrow_tx_hash": {
            "type": [
              "string",
              "null"
            ],
            "nullable": true,
            "example": "abc123txhash"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateTradeBody": {
        "type": "object",
        "required": [
          "assetType",
          "amount",
          "expiresInHours"
        ],
        "properties": {
          "assetType": {
            "type": "string",
            "maxLength": 50,
            "pattern": "^[A-Za-z0-9_-]+$",
            "example": "MTN_AIRTIME"
          },
          "amount": {
            "type": "number",
            "exclusiveMinimum": 0,
            "example": 500
          },
          "expiresInHours": {
            "type": "integer",
            "minimum": 1,
            "maximum": 168,
            "example": 24
          }
        }
      },
      "BuyTradeBody": {
        "type": "object",
        "required": [
          "signedXdr"
        ],
        "properties": {
          "signedXdr": {
            "type": "string",
            "description": "The escrow deposit envelope returned by `POST /api/v1/trades/{id}/buy/prepare`, signed by the buyer in their own browser. Base64 XDR. The buyer's secret key is never sent to this API.",
            "example": "AAAAAgAAAAD..."
          }
        }
      },
      "PrepareBuyResponse": {
        "type": "object",
        "required": [
          "xdr",
          "networkPassphrase",
          "publicKey"
        ],
        "properties": {
          "xdr": {
            "type": "string",
            "description": "Unsigned, already-simulated transaction envelope (base64 XDR)."
          },
          "networkPassphrase": {
            "type": "string",
            "description": "Network the envelope must be signed against.",
            "example": "Test SDF Network ; September 2015"
          },
          "publicKey": {
            "type": "string",
            "description": "Buyer's Stellar account — the envelope's source account."
          }
        }
      },
      "TradeListResponse": {
        "type": "object",
        "required": [
          "data",
          "pagination"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TradeOffer"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        }
      },
...
      "ProfileTradesResponse": {
        "type": "object",
        "required": [
          "data",
          "pagination"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TradeOffer"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid Bearer JWT.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Missing or invalid Authorization header"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "The requested resource does not exist.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Trade offer not found"
            }
          }
        }
      },
      "UnprocessableEntity": {
        "description": "Request body failed schema validation.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationErrorResponse"
            }
          }
        }
      },
...
      "InternalError": {
        "description": "Unexpected server error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": "Internal server error"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/trades": {
      "get": {
        "tags": [
          "Trades"
        ],
        "summary": "List active trade offers",
        "description": "Returns paginated active listings that have not expired.",
        "security": [],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of active trade offers.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradeListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "post": {
        "tags": [
          "Trades"
        ],
        "summary": "Create a trade offer",
        "description": "Creates a new trade listing. Calls `create_listing` on the Soroban escrow contract and stores the resulting on-chain listing ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTradeBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Trade offer created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/TradeOffer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Seller wallet not found or contract call failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/v1/trades/{id}": {
      "get": {
        "tags": [
          "Trades"
        ],
        "summary": "Get trade by ID",
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Trade offer found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/TradeOffer"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/v1/trades/{id}/buy/prepare": {
      "post": {
        "tags": [
          "Trades"
        ],
        "summary": "Build the unsigned escrow deposit for a buy",
        "description": "Returns the simulated but unsigned `deposit_to_escrow` envelope for the authenticated buyer to sign locally. Takes no request body. Sign the returned XDR client-side, then POST it to `/api/v1/trades/{id}/buy`.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Unsigned envelope ready for client-side signing.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PrepareBuyResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Trade not available, already locked, or buyer is the seller.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/api/v1/trades/{id}/buy": {
      "post": {
        "tags": [
          "Trades"
        ],
        "summary": "Submit a signed escrow deposit for a trade",
        "description": "Submits the buyer-signed `deposit_to_escrow` envelope from `/buy/prepare` to the network. Trade status transitions to Locked.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuyTradeBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Escrow deposit confirmed — trade is now Locked.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/TradeOffer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Trade not available, already locked, or buyer is the seller.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
    "/api/v1/profile/trades": {
      "get": {
        "tags": [
          "Profile"
        ],
        "summary": "Get own trade history",
        "description": "Returns paginated trade history for the authenticated user (as seller or buyer). Filterable by status.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 10
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "All",
                "Active",
                "Locked",
                "Completed",
                "Cancelled",
                "Disputed"
              ],
              "default": "All"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated trade history.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileTradesResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
...
}